refactor : 프로젝트 리스트 토글 부분 코드 분리
